+++ test/integration/tests/4938-non-ascii-module-names/files/Lib.hs
@@ -1,0 +1,9 @@
+module Lib
+    ( foo
+    , bar
+    ) where
+
+import Täst
+
+bar :: Int
+bar = 42